Code patching system; Block & Collectible patches
